requirements.txt
paho-mqtt<2.0.0
python-dotenv
requests
pyyaml